--- ghost/slack-notifications/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/slack-notifications');
