--- ghost/slack-notifications/lib/slack-notifications.js
@@ -1,2 +1,0 @@
-module.exports.SlackNotificationsService = require('./SlackNotificationsService');
-module.exports.SlackNotifications = require('./SlackNotifications');
