--- ghost/slack-notifications/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@tryghost/slack-notifications",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/slack-notifications",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing TZ=UTC c8 --all --check-coverage --100  --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.8.2",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/validator": "0.2.14",
-    "@tryghost/version": "0.1.30",
-    "got": "9.6.0"
-  }
-}
